less-stache: only expand mustache fields on start/stop cmds and only if needed
